# OpenAPI spec.yml
openapi: 3.0.1
info:
  title: Web Market
  version: "0.1"
externalDocs:
  url: https://github.com/lucianoScotellaro/SWA-WebMarketRESTfulAPI.git
  description: GitHub repository of the project
servers:
  - url: http://localhost:8080/webmarket/rest
tags:
  - name: Request
    description: Operations on a single request
  - name: Requests Collection
    description: Operations on requests collection
  - name: Request property
    description: Operations on a single request properties
  - name: Auth
    description: Authentication operations
paths:
  /auth/login:
    post:
      operationId: login
      description: Login
      tags:
        - Auth
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Password"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "500":
          $ref: "#/components/responses/GeneralError"
  /auth/logout:
    delete:
      operationId: logout
      description: Logout
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests:
    get:
      operationId: getTechnicianRequests
      description: Returns requests handled by the technician
      tags:
        - Requests Collection
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TechnicianID"
      responses:
        "200":
          description: Technician requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianBaseRequestsList"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
    post:
      operationId: addRequest
      description: add new request
      tags:
        - Request
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Request"
      responses:
        "201":
          $ref: "#/components/responses/SuccessfulInsert"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests/ongoing:
    get:
      operationId: getOrdererOngoingRequests
      description: Returns ongoing requests for an orderer
      tags:
        - Requests Collection
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrdererID"
      responses:
        "200":
          description: Orderer open requests base info list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdererBaseRequestsList"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests/unassigned:
    get:
      operationId: getUnassignedRequests
      description: Returns requests that have not been assigned to a technician
      tags:
        - Requests Collection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unassigned requests base info list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnassignedRequestsList"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests/{requestid}:
    get:
      operationId: getRequest
      description: Returns a request
      tags:
        - Request
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
    delete:
      operationId: deleteRequest
      description: Delete a request
      tags:
        - Request
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests/{requestid}/technician:
    put:
      operationId: setTechnician
      description: Binds a technician to a request
      tags:
        - Request
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests/{requestid}/proposal:
    get:
      operationId: getProposal
      description: Returns proposal associated to the request
      tags:
        - Request
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      responses:
        "200":
          description: Proposal associated to the request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
    put:
      operationId: setProposal
      description: Add or update a proposal for a request
      tags:
        - Request property
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Proposal"
      responses:
        "201":
          $ref: "#/components/responses/SuccessfulInsert"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
  /requests/{requestid}/proposal/answer:
    put:
      operationId: setAnswer
      description: Add answer for a proposal
      tags:
        - Request property
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Answer"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/GeneralError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
  schemas:
    ID:
      type: integer
      minimum: 1
    Email:
      type: string
      format: email
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        email:
          $ref: "#/components/schemas/Email"
        username:
          type: string
      required:
        - id
        - email
        - username
      additionalProperties: false
    Product:
      type: object
      properties:
        name:
          type: string
        producer:
          type: string
        id:
          $ref: "#/components/schemas/ID"
        price:
          type: number
          minimum: 0
        url:
          type: string
      required:
        - name
        - producer
        - id
        - price
        - url
      additionalProperties: false
    Proposal:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        notes:
          type: string
        answer:
          $ref: "#/components/schemas/Answer"
      required:
        - product
        - notes
        - answer
      additionalProperties: false
    Request:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        orderer:
          $ref: "#/components/schemas/User"
        technician:
          $ref: "#/components/schemas/User"
        category:
          type: string
        characteristics:
          type: object
        notes:
          type: string
        proposal:
          $ref: "#/components/schemas/Proposal"
        ongoing:
          type: boolean
        outcome:
          type: string
          enum:
            - Accepted
            - Rejected as non compliant
            - Rejected as defective
      required:
        - id
        - orderer
        - technician
        - category
        - characteristics
        - notes
        - proposal
        - ongoing
        - outcome
      additionalProperties: false
    OrdererBaseRequestsList:
      type: array
      items:
        type: object
        properties:
          id:
            $ref: "#/components/schemas/ID"
          technician:
            $ref: "#/components/schemas/User"
          url:
            type: string
        required:
          - id
          - technician
          - url
    UnassignedRequestsList:
      type: array
      items:
        type: object
        properties:
          id:
            $ref: "#/components/schemas/ID"
          orderer:
            $ref: "#/components/schemas/User"
          url:
            type: string
        required:
          - id
          - orderer
          - url
    TechnicianBaseRequestsList:
      type: array
      items:
        type: object
        properties:
          id:
            $ref: "#/components/schemas/ID"
          orderer:
            $ref: "#/components/schemas/User"
          ongoing:
            type: boolean
          url:
            type: string
        required:
          - id
          - orderer
          - ongoing
          - url
    Answer:
      type: object
      properties:
        accepted:
          type: boolean
        reason:
          type: string
      required:
        - accepted
        - reason
      additionalProperties: false
  parameters:
    RequestID:
      name: requestid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"
    ProposalID:
      name: proposalid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"
    OrdererID:
      name: ordererid
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/ID"
    TechnicianID:
      name: techid
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/ID"
    Username:
      name: username
      in: query
      required: true
      schema:
        type: string
    Password:
      name: password
      in: query
      required: true
      schema:
        type: string
  responses:
    SuccessfulInsert:
      description: Element created
      headers:
        location:
          schema:
            type: string
            format: uri
      content:
        application/json:
          schema:
            type: string
            format: uri
    NoContent:
      description: No content
    GeneralError:
      description: General error
    UnauthorizedError:
      description: Missing or invalid bearer token
    NotFound:
      description: Element not found
      content:
        text/plain:
          schema:
            type: string
            example: Not found